requirements.txt
coverage
django
django-allauth
django-allauth-ui
django-debug-toolbar
django-fastdev
django-filter
django-guardian
django-oscar
django-redis
djlint
hiredis
isort
nplusone
parameterized
pillow
pip
pinax-messages
psycopg[c]
psycopg_pool
pycountry
python-dotenv
python-telegram-handler
redis
ruff
sorl-thumbnail
uv
wheel